Add a description of a recommended publication workflow
This adds the recommendation that the "name" attribute of the top
level object in the descriptor should be unique, preferably
globally but at least unique within the registry of the budget data
package.

This also adds a section about a recommended way of constructing
budget data packages, what resources should be included in a single
budget data package and what resources should be split across
multiple budget data packages.