Name all of the properties that a dimension can include
The two dimension properties are fields, primaryKey and name
(required fields) and then dimensionType (a MAY). Name the
different dimension types as part of the explanation of
dimensionType instead of just a list that appears all of a sudden.

This requires a minor reconstruction of the dimension section:
Removing redundant examples and structure the properties list
similarly to the measures.