Restructure primaryKey to conform with JSON Table Schema
Instead of having primaryKey as an empty string or a boolean on a
particular field, the primary key should be a property in and of
itself. This separates concerns within the dimension (primary key
is about the dimension not the field declaration), makes it
correspond to the JSON Table Schema convention (why do it in a
different way?), makes it easier to access and makes it possible
to create a schema.