Restructure mapping to be a hash with measures and dimensions properties
Instead of having properties with measure and dimension names as
keys, have measures and dimensions of the mapping hash to be arrays
of hashes, each hash contains a property called "name" with a value
representing the measure or dimension name. This conforms to data
package conventions and makes creating schemas easier but does make
it harder to access the right measure or the right dimension as one
will have to traverse through the array. However, since the measure
and dimension names can be just about anything that shouldn't
really matter.